# swagger/SocialnetApiDocs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Socialnet API",
    "description": "Code Lounge",
    "version": "1.6"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/post": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "summary": "Получение списка постов",
        "operationId": "getPost",
        "parameters": [
          {
            "name": "withFriends",
            "in": "query",
            "description": "Флаг \"с друзьями\"",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Сортировка",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "Удаленные посты включены",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "accountIds",
            "in": "query",
            "description": "ID аккаунта",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Теги",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "dateForm",
            "in": "query",
            "description": "Дата от",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "description": "Дата до",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "description": "Автор",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "text",
            "in": "query",
            "description": "Текст",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Посты успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoPostDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "post-controller"
        ],
        "summary": "Обновление поста",
        "operationId": "getUpdatePost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Пост успешно обновлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "Создание поста",
        "operationId": "getCreatePost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Пост успешно создан",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/notifications/settings": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Получение настроек уведомлений",
        "operationId": "getNotificationSetting",
        "responses": {
          "200": {
            "description": "Настройки уведомлений успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Обновление настроек уведомлений",
        "operationId": "updateNotificationSetting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Настройки уведомлений успешно обновлены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/notifications/readed": {
      "put": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Отметить все уведомления как прочитанные",
        "operationId": "readAllNotifications",
        "responses": {
          "200": {
            "description": "Уведомления отмечены как прочитанные",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/post/{postId}/like": {
      "post": {
        "tags": [
          "like-controller"
        ],
        "summary": "Получение лайка по посту",
        "operationId": "getLike",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLikeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Лайк успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LikeDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "like-controller"
        ],
        "summary": "Удаление лайка по посту",
        "operationId": "deleteLike",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Лайк успешно удален",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/post/{postId}/comment/{commentId}/like": {
      "post": {
        "tags": [
          "like-controller"
        ],
        "summary": "Получение лайка по комментарию",
        "operationId": "getLikeByComment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "ID комментария",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Лайк успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LikeDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "like-controller"
        ],
        "summary": "Удаление лайка по комментарию",
        "operationId": "deleteLikeByComment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID поста",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "description": "ID комментария",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Лайк успешно удален",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/notifications/settings/{id}": {
      "post": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Установка настроек уведомлений",
        "operationId": "setNotificationSetting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Настройки уведомлений успешно установлены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Получение уведомлений",
        "operationId": "getNotifications",
        "responses": {
          "200": {
            "description": "Уведомления успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoDataTimestampDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/notifications/count": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Получение количества уведомлений",
        "operationId": "getNotificationsCount",
        "responses": {
          "200": {
            "description": "Количество уведомлений успешно получено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DataTimestampDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор поста",
            "format": "uuid"
          },
          "imagePath": {
            "type": "string",
            "description": "Путь к изображению"
          },
          "postText": {
            "type": "string",
            "description": "Текст поста"
          },
          "publishDate": {
            "type": "string",
            "description": "Дата публикации"
          },
          "tags": {
            "type": "array",
            "description": "Теги",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "title": {
            "type": "string",
            "description": "Заголовок"
          }
        },
        "description": "Данные для создания поста"
      },
      "TagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID тега",
            "format": "int64"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Флаг удаления"
          },
          "name": {
            "type": "string",
            "description": "Имя тега"
          }
        },
        "description": "Объект тега"
      },
      "PostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор поста",
            "format": "uuid"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Флаг удаления поста"
          },
          "time": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Время создания поста"
          },
          "timeChanged": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Время изменения поста"
          },
          "authorId": {
            "type": "string",
            "description": "Идентификатор автора поста",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Заголовок поста"
          },
          "type": {
            "type": "string",
            "description": "Тип поста",
            "enum": [
              "POSTED",
              "QUEUED"
            ]
          },
          "postText": {
            "type": "string",
            "description": "Текст поста"
          },
          "isBlocked": {
            "type": "boolean",
            "description": "Флаг блокировки поста"
          },
          "commentsCount": {
            "type": "integer",
            "description": "Количество комментариев",
            "format": "int32"
          },
          "reactions": {
            "type": "array",
            "description": "Реакции на пост",
            "items": {
              "$ref": "#/components/schemas/ReactionDto"
            }
          },
          "myReaction": {
            "type": "string",
            "description": "Моя реакция на пост"
          },
          "tags": {
            "type": "array",
            "description": "Теги поста",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "likeAmount": {
            "type": "integer",
            "description": "Количество лайков",
            "format": "int32"
          },
          "myLike": {
            "type": "boolean",
            "description": "Мой лайк"
          },
          "imagePath": {
            "type": "string",
            "description": "Путь к изображению"
          },
          "publishDate": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Дата публикации поста"
          }
        },
        "description": "Данные поста"
      },
      "ReactionDto": {
        "type": "object",
        "properties": {
          "reactionType": {
            "type": "string",
            "description": "Тип реакции"
          },
          "count": {
            "type": "integer",
            "description": "Количество",
            "format": "int32"
          }
        },
        "description": "Данные реакции"
      },
...
      "NotificationUpdateDto": {
        "type": "object",
        "properties": {
          "notificationType": {
            "type": "string",
            "description": "Тип уведомления",
            "enum": [
              "LIKE",
              "POST",
              "POST_COMMENT",
              "COMMENT_COMMENT",
              "MESSAGE",
              "FRIEND_REQUEST",
              "FRIEND_BIRTHDAY",
              "SEND_EMAIL_MESSAGE"
            ]
          },
          "enable": {
            "type": "boolean",
            "description": "Включено ли"
          }
        },
        "description": "Данные для обновления настроек уведомлений"
      },
      "NotificationSettingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID настройки",
            "format": "uuid"
          },
          "enablePost": {
            "type": "boolean",
            "description": "Уведомления о публикациях"
          },
          "enablePostComment": {
            "type": "boolean",
            "description": "Уведомления о комментариях к публикациям"
          },
          "enableCommentComment": {
            "type": "boolean",
            "description": "Уведомления о комментариях к комментариям"
          },
          "enableMessage": {
            "type": "boolean",
            "description": "Уведомления о сообщениях"
          },
          "enableFriendRequest": {
            "type": "boolean",
            "description": "Уведомления о запросах в друзья"
          },
          "enableFriendBirthday": {
            "type": "boolean",
            "description": "Уведомления о днях рождения друзей"
          },
          "enableSendEmailMessage": {
            "type": "boolean",
            "description": "Отправка уведомлений по электронной почте"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Удалена ли"
          }
        },
        "description": "Настройки уведомлений"
      },
...
      "CreateLikeDto": {
        "type": "object",
        "properties": {
          "reactionType": {
            "type": "string",
            "description": "Тип реакции"
          },
          "type": {
            "type": "string",
            "description": "Тип публикации",
            "enum": [
              "COMMENT",
              "POST"
            ]
          }
        },
        "description": "Данные для создания лайка"
      },
      "LikeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID лайка",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "description": "ID автора",
            "format": "uuid"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Удален ли лайк"
          },
          "itemId": {
            "type": "string",
            "description": "ID объекта лайка",
            "format": "uuid"
          },
          "reactionType": {
            "type": "string",
            "description": "Тип реакции"
          },
          "time": {
            "type": "string",
            "description": "Время создания",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "description": "Тип публикации",
            "enum": [
              "COMMENT",
              "POST"
            ]
          }
        },
        "description": "Данные лайка"
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageResponseDtoPostDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "description": "Общее количество элементов",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "Общее количество страниц",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "first": {
            "type": "boolean",
            "description": "Первая страница"
          },
          "last": {
            "type": "boolean",
            "description": "Последняя страница"
          },
          "numberOfElements": {
            "type": "integer",
            "description": "Количество элементов на странице",
            "format": "int32"
          },
          "empty": {
            "type": "boolean",
            "description": "Пустая страница"
          },
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableDto"
          },
          "content": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "Содержимое страницы",
            "items": {
              "$ref": "#/components/schemas/PostDto"
            }
          }
        },
        "description": "Ответ с данными страницы"
      },
      "PageableDto": {
        "type": "object",
        "properties": {
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageNumber": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "description": "Смещение",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean",
            "description": "Без разбивки на страницы"
          },
          "paged": {
            "type": "boolean",
            "description": "Разбито на страницы"
          }
        },
        "description": "Данные для постраничного вывода"
      },
      "SortDto": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean",
            "description": "Не отсортировано"
          },
          "sorted": {
            "type": "boolean",
            "description": "Отсортировано"
          },
          "empty": {
            "type": "boolean",
            "description": "Пусто"
          }
        },
        "description": "Данные сортировки"
      },
...
      "DataTimestampDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Временна́я метка",
            "format": "date-time"
          },
          "data": {
            "type": "object",
            "description": "Данные"
          }
        },
        "description": "Данные для уведомления с timestamp"
      },
      "PageResponseDtoDataTimestampDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "description": "Общее количество элементов",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "Общее количество страниц",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "first": {
            "type": "boolean",
            "description": "Первая страница"
          },
          "last": {
            "type": "boolean",
            "description": "Последняя страница"
          },
          "numberOfElements": {
            "type": "integer",
            "description": "Количество элементов на странице",
            "format": "int32"
          },
          "empty": {
            "type": "boolean",
            "description": "Пустая страница"
          },
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableDto"
          },
          "content": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "Содержимое страницы",
            "items": {
              "$ref": "#/components/schemas/DataTimestampDto"
            }
          }
        },
        "description": "Ответ с данными страницы"
      },
...
    }
  }
}